fix: use compose-go to resolve extends in PullableServices
PullableServices used custom YAML parsing that didn't handle the
extends directive, causing extended services to be incorrectly
included in compose pull. Delegate to compose-go's ReadProject
which fully resolves extends (including cross-file references).
